requirements.txt
scikit-learn
umap-learn
prince
matplotlib
Flask

